Move 'node' into 'nodes' subfolder.
