+++ earthmover/nodes/transformation.py
@@ -1,4 +1,4 @@
-from earthmover.node import Node
+from earthmover.nodes.node import Node
 from earthmover.operations.operation import Operation
 
 from typing import List, Tuple
